chore: 테스트 용 예제 코드 이동
